# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexus CRM API",
    "version": "1.0.0",
    "description": "REST API for the Nexus CRM application.\n\n## Authentication\n\nTwo auth methods are supported:\n\n- **BearerAuth** \u2013 `Authorization: Bearer <token>`. Per-user API token generated in the dashboard. Scoped to the token owner's data (admin tokens get cross-tenant read access).\n- **CookieAuth** \u2013 Session cookie set by the Better Auth `/api/auth` flow (Google OAuth).\n\nToken management (POST/DELETE `/api/token`) requires a session cookie. File downloads require CookieAuth or BearerAuth; unauthenticated sharing uses per-link `/s/{code}` URLs scoped to the link owner."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current host"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Per-user API token generated via the dashboard (POST /api/token). Scoped to the token owner's data. Admin users' tokens get cross-tenant read access."
      },
      "CookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "better-auth.session_token",
        "description": "Session cookie issued by the Better Auth flow at `/api/auth`. Admin users (isAdmin=true) bypass per-user filtering."
      }
    },
    "schemas": {
      "ApplicationStatus": {
        "type": "string",
        "enum": ["inbound", "applied", "interview", "offer", "rejected"],
        "description": "Current stage of the opportunity pipeline."
      },
      "ContactRef": {
        "type": "object",
        "description": "A contact person associated with an application.",
        "required": ["id", "name", "applicationId", "createdAt"],
        "properties": {
          "id":            { "type": "integer", "example": 1 },
          "name":          { "type": "string",  "example": "Jane Smith" },
          "email":         { "type": ["string", "null"], "format": "email", "example": "jane@example.com" },
          "phone":         { "type": ["string", "null"], "example": "+1-555-0100" },
          "role":          { "type": ["string", "null"], "example": "Engineering Manager" },
          "linkedIn":      { "type": ["string", "null"], "example": "https://linkedin.com/in/janesmith" },
          "applicationId": { "type": "integer", "example": 42 },
          "createdAt":     { "type": "string",  "format": "date-time" }
        }
      },
...
      "Application": {
        "type": "object",
        "description": "An opportunity record.",
        "required": ["id", "userId", "company", "role", "status", "createdAt", "updatedAt", "contacts"],
        "properties": {
          "id":             { "type": "integer",              "example": 42 },
          "userId":         { "type": "string",               "example": "user_abc123" },
          "company":        { "type": "string",               "example": "Acme Corp" },
          "role":           { "type": "string",               "example": "Senior Engineer" },
          "status":         { "$ref": "#/components/schemas/ApplicationStatus" },
          "appliedAt":      { "type": ["string", "null"],     "format": "date-time" },
          "lastContact":    { "type": ["string", "null"],     "format": "date-time" },
          "followUpAt":     { "type": ["string", "null"],     "format": "date-time" },
          "notes":          { "type": ["string", "null"],     "example": "Had a great first call." },
          "jobDescription": { "type": ["string", "null"],     "example": "We are looking for…" },
          "canonicalJobUrl": { "type": ["string", "null"], "format": "uri" },
          "workMode": { "type": ["string", "null"], "enum": ["remote", "hybrid", "onsite", "flexible", null] },
          "eligibleCountries": { "type": "array", "items": { "type": "string" } },
          "primaryLocations": { "type": "array", "items": { "type": "string" } },
          "officeDaysMin": { "type": ["integer", "null"] },
          "travelPercent": { "type": ["integer", "null"] },
          "visaSponsorship": { "type": ["boolean", "null"] },
          "rightToWorkRequired": { "type": ["boolean", "null"] },
          "timezoneOverlap": { "type": ["string", "null"] },
          "salaryCurrency": { "type": ["string", "null"] },
          "salaryPeriod": { "type": ["string", "null"] },
          "salaryType": { "type": ["string", "null"] },
          "atsName": { "type": ["string", "null"] },
          "requisitionId": { "type": ["string", "null"] },
          "jobCapturedAt": { "type": ["string", "null"], "format": "date-time" },
          "jobVerifiedAt": { "type": ["string", "null"], "format": "date-time" },
          "jobPostedAt": { "type": ["string", "null"], "format": "date-time" },
          "jobClosedAt": { "type": ["string", "null"], "format": "date-time" },
          "jobContentHash": { "type": ["string", "null"] },
          "jobLiveness": { "type": ["string", "null"] },
          "jobSummary": { "type": ["string", "null"] },
          "currentStage": { "type": ["string", "null"] },
          "createdAt":      { "type": "string",               "format": "date-time" },
          "updatedAt":      { "type": "string",               "format": "date-time" },
          "contacts":       {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ContactRef" }
          }
        }
      },
...
      "ApplicationEvent": {
        "type": "object",
        "required": ["id", "applicationId", "type", "occurredAt"],
        "properties": {
          "id": { "type": "string" },
          "applicationId": { "type": "string" },
          "type": { "type": "string" },
          "idempotencyKey": { "type": ["string", "null"] },
          "occurredAt": { "type": "string", "format": "date-time" },
          "source": { "type": ["string", "null"] },
          "actor": { "type": ["string", "null"] },
          "metadata": { "type": ["object", "null"], "additionalProperties": true }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "example": "Unauthorized" }
        }
      },
      "SuccessResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": { "type": "boolean", "example": true }
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing or invalid credentials.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Unauthorized" }
          }
        }
      },
      "Forbidden": {
        "description": "Insufficient permissions — admin privileges required or resource does not belong to the authenticated user.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Session required" }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found or does not belong to the authenticated user.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" },
            "example": { "error": "Not found" }
          }
        }
      },
      "BadRequest": {
        "description": "Invalid or missing request parameters.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/applications/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "schema": { "type": "integer", "minimum": 1 },
          "description": "Application ID."
        }
      ],
      "get": {
        "operationId": "getApplication",
        "summary": "Get application",
        "description": "Returns a single application by ID. Access is scoped to the authenticated user (admin users see all).",
        "tags": ["Applications"],
        "security": [
          { "BearerAuth": [] },
          { "CookieAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "The application.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Application" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "operationId": "updateApplication",
        "summary": "Update application",
        "description": "Partially updates an application. All fields are optional. **Requires a session.**",
        "tags": ["Applications"],
        "security": [
          { "BearerAuth": [] },
          { "CookieAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "company":        { "type": "string", "maxLength": 255 },
                  "role":           { "type": "string", "maxLength": 255 },
                  "status":         { "$ref": "#/components/schemas/ApplicationStatus" },
                  "appliedAt":      { "type": ["string", "null"], "format": "date-time" },
                  "lastContact":    { "type": ["string", "null"], "format": "date-time" },
                  "followUpAt":     { "type": ["string", "null"], "format": "date-time" },
                  "notes":          { "type": ["string", "null"], "maxLength": 10000 },
                  "jobDescription": { "type": ["string", "null"], "maxLength": 50000 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated application.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Application" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "operationId": "deleteApplication",
        "summary": "Delete application",
        "description": "Deletes an application and its related contacts. **Requires a session.**",
        "tags": ["Applications"],
        "security": [
          { "BearerAuth": [] },
          { "CookieAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Deletion confirmed.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/api/applications/{id}/events": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "integer", "minimum": 1 } }
      ],
      "get": {
        "operationId": "listApplicationEvents",
        "summary": "List append-only application timeline events",
        "tags": ["Applications"],
        "security": [{ "BearerAuth": [] }, { "CookieAuth": [] }],
        "responses": {
          "200": {
            "description": "Timeline events.",
            "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ApplicationEvent" } } } }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "post": {
        "operationId": "recordApplicationEvent",
        "summary": "Append an application timeline event",
        "tags": ["Applications"],
        "security": [{ "BearerAuth": [] }, { "CookieAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["type"],
                "properties": {
                  "type": { "type": "string", "maxLength": 100 },
                  "idempotencyKey": { "type": ["string", "null"] },
                  "occurredAt": { "type": "string", "format": "date-time" },
                  "metadata": { "type": "object", "additionalProperties": true }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created timeline event." },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
...
}